# server/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VPS CI/CD Automation Runner API",
    "description": "Webhook-driven git sync, automated zero-downtime deployment runner, and telemetry platform for self-hosted VPS instances.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Server"
    }
  ],
...
  "paths": {
...
    "/api/services": {
      "get": {
        "tags": [
          "Services"
        ],
        "summary": "List all configured sync services",
        "responses": {
          "200": {
            "description": "List of services"
          }
        }
      },
      "post": {
        "tags": [
          "Services"
        ],
        "summary": "Create a new sync service",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service created"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/api/services/{id}": {
      "get": {
        "tags": [
          "Services"
        ],
        "summary": "Get service details by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service details"
          },
          "404": {
            "description": "Service not found"
          }
        }
      },
      "put": {
        "tags": [
          "Services"
        ],
        "summary": "Update service configuration",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Service updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Services"
        ],
        "summary": "Delete service",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service deleted"
          }
        }
      }
    },
    "/api/services/{id}/sync": {
      "post": {
        "tags": [
          "Services"
        ],
        "summary": "Manually trigger immediate git sync and deployment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "branch": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Sync queued"
          }
        }
      }
    },
    "/api/services/{id}/rollback": {
      "post": {
        "tags": [
          "Services"
        ],
        "summary": "Rollback service to a target commit SHA",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "targetSha"
                ],
                "properties": {
                  "targetSha": {
                    "type": "string"
                  },
                  "branch": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Rollback queued"
          }
        }
      }
    },
    "/api/services/{id}/commits": {
      "get": {
        "tags": [
          "Services"
        ],
        "summary": "Get git commit history for service repository",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Commit log"
          }
        }
      }
    },
    "/api/services/{id}/env": {
      "get": {
        "tags": [
          "Services"
        ],
        "summary": "Get encrypted environment variables keys for service",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Environment variables"
          }
        }
      },
      "put": {
        "tags": [
          "Services"
        ],
        "summary": "Save encrypted environment variables for service",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "env": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "key": {
                          "type": "string"
                        },
                        "value": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Environment variables updated"
          }
        }
      }
    },
...
    "/api/services/{id}/notifications": {
      "get": {
        "tags": [
          "Services"
        ],
        "summary": "List notification bindings for a service",
        "responses": {
          "200": {
            "description": "Bindings list"
          }
        }
      },
      "put": {
        "tags": [
          "Services"
        ],
        "summary": "Replace notification bindings for a service",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bindings": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "channel_id": {
                          "type": "integer"
                        },
                        "on_start": {
                          "type": "boolean"
                        },
                        "on_success": {
                          "type": "boolean"
                        },
                        "on_failure": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bindings saved"
          }
        }
      }
    },
...
    "/api/services/{id}/triggers": {
      "get": {
        "tags": [
          "Services"
        ],
        "summary": "List triggers for a service",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 30
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trigger list"
          },
          "404": {
            "description": "Service not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ServiceInput": {
        "type": "object",
        "required": [
          "name",
          "repo_url",
          "folder_path"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "provider": {
            "type": "string",
            "enum": [
              "github",
              "gitlab",
              "bitbucket",
              "gitea",
              "gogs",
              "generic"
            ]
          },
          "repo_url": {
            "type": "string"
          },
          "folder_path": {
            "type": "string"
          },
          "branch_mode": {
            "type": "string",
            "enum": [
              "webhook",
              "fixed",
              "current"
            ]
          },
          "fixed_branch": {
            "type": "string"
          },
          "allowed_branches": {
            "type": "string"
          },
          "sync_mode": {
            "type": "string",
            "enum": [
              "pull",
              "reset"
            ]
          },
          "healthcheck_url": {
            "type": "string"
          },
          "auto_rollback": {
            "type": "boolean"
          },
          "commands": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "command": {
                  "type": "string"
                },
                "branch_filter": {
                  "type": "string"
                },
                "continue_on_error": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    }
  }
}